Mostly wrapped up pool docs
